made user data fetch only once and used as props everywhere
