+++ Frontend/src/pages/home.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Canvas } from "@react-three/fiber";
 // import "bootstrap/dist/css/bootstrap.min.css";
 
-import PNavbar from "../components/navbar.jsx";
+// import PNavbar from "../components/navbar.jsx";
 import Splinemodel from "../components/spline.jsx";
 import "../style.css";
 import About from "../components/about.jsx";
@@ -10,7 +10,7 @@ import About from "../components/about.jsx";
 export default function Home() {
   return (
     <>
-      <PNavbar />
+      {/* <PNavbar /> */}
       <div className="homepage">
         <Splinemodel />
       </div>
